Restructure io module, again
